Add deprecated_group Opt kwarg
We often wish to take existing configuration options and move them
from the DEFAULT group to a more specialized group, while still
supporting existing configurations.

We can support this by adding a new deprecated_group kwarg which
allows an option to be aliased to a name in another group e.g.

  CONF.register_opt(cfg.StrOpt('foo',
                               deprecated_name='bar',
                               deprecated_group='DEFAULT'),
                    group='blaa')

or where the option was simply moved between groups:

  CONF.register_opt(cfg.StrOpt('foo',
                               deprecated_group='DEFAULT'),
                    group='blaa')

Implements blueprint cfg-move-opts-between-groups

Change-Id: I4d974c71fab3e3c5c7a3e6514bdd9cb20e7200ff